Fix segfault in Authenticode hash check (#16042)
I noticed that `r2` will crash when loading a PE file with
Authenticode digest algorithm other than SHA-1 or MD5. I traced
it down to the `PE_(bin_pe_compute_authentihash)` function returning
`NULL` if it encounters an unsupported digest function. This results
in `NULL` being passed to `strcmp` which causes the segfault.

Solution was to add a check for `PE_(bin_pe_compute_authentihash)`
returning `NULL` and to set `bin->is_authhash_valid` to `NULL`.
The real solution is to add support for more algorithms but this will
stop crashes for now.