format/elf: switch from _offset functions to _addr functions
It is not always possible to do the following transformations:
1) vaddr -> paddr -> vaddr
2) paddr -> vaddr -> paddr

Given this, it is better to directly return a full RBinAddr structure
instead of converting v2p and p2v across callee and callers, risking to
get the wrong addresses during those unneeded transformations.

Moreover, find an entrypoint only when one is specified, otherwise do
not fall back to assuming the first byte of a .text/.init.text sections
are the entry point.