Make dyldcache accelerator info optional ##bin
This allows to open dyld cache files from iOS 13.4 for which
apparently the accelerator info are missing.

What’s lost when this info is missing is the ability to efficiently
detect which single library depends on which other libs when
performing the filtering.

That means that for iOS 13.4 library caches, the user has to
manually specify all the dependencies in the filter.