Support arbitrary length identifiers in Mach-O ##bin (#17064)
* Convert symbol_t name member from char[R_BIN_MACH0_STRING_LENGTH] to char *

* Invert the variables that assmebles the ht key to ensure uniqueness

Long identiers will fill the key buffer with the first 256 chars of the
name and have the rest of the name and the addresss ignored. This can
lead to two different identifiers having the same key.

The first 256 characters of two different identifiers might be
identical, while their address won't. Therefore, to ensure different
identifiers will have different keys, we sholud first emmbed the part
that is ought to be unique in the identifier, the address, and only then
take the identifier itself into account.

* Add testcase

* Define variable outside of the for-loop block

* Remove redundant conditional free

* Replace sdb_fmt with r_str_newf

* Add missing format string argument and use r_str_newf

* Use NULL instead of 0 for name pointer

* Change last to bool and adjust rvalue accordingly

* Ensure we do not process empty an empty string

Co-authored-by: phakeobj <phakeobj@users.noreply.github.com>