Do not return a buffer when slurping a empty file
While running r2 from git, I was greeted by:

   |ERROR| Invalid command ��Y�' (0x08)

I quickly found this came from:

   $ ls -l ~/.radare2rc
   -rw-r--r--. 1 misc misc 0 19 avril  2017 /home/misc/.radare2rc

Debugging showed that if the file is empty, we return the buffer
that was reallocate earlier with uninitialized memory without adding
a NULL byte.

Since r_core_cmd_file do not check the size (it pass NULL as a
2nd parameter for r_file_slurp), it doesn't see the file is empty,
and so try to interpret 6 bytes as a command.

Given that grep find 54 calls to r_file_slurp with NULL
as a 2nd parameter, this will also fix a bit more than just this
error message.