fortunes: remove nsfw and creepy fortunes
These fortunes make a lot of people unconfortable and can be trigering
for some. Even if they are not active by default, they should have no
place here if you are trying to be a welcoming project.

There are a lot of other ways to be quirky and fun, I see no logical
reason to have these fortunes knowing they will negatively impact
people.

Signed-off-by: Filipe Laíns <lains@riseup.net>