Fix a credentials downgrade problem in Install.sh ##build
* Fix failure when finding the executing user's ID during install

The `UID` variable used previously is a Bash extension that isn't present on all
POSIX shells.

* Prioritize downgrading credentials during install

The first thing to do is to check and, if necessary, downgrade the credentials
before proceeding any further during install. Also, checking for the presence of
GNU Make is delayed after the arguments to the install script have been parsed.