Allow to override build date with SOURCE_DATE_EPOCH
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.
This date call only works with GNU date.

Also use UTC to be independent of timezone.

This is the equivalent of 6b260b87c3345568ebeddf57fbe95c864ee8baf2 for meson.

This PR was done while working on reproducible builds for openSUSE.