In cBPF jt and jf are unsigned, fix the code ##anal
The cBPF architecture defines that conditional operators offsets (jt and
jf) are unsigned, where 0 means the next operator, this way the code
flow is always loop-free.  However, the C code treated jt and jf as
signed, which for large enough values made the offset negative and broke
a number of things.

For example, in the bytecode compiled from the filter expression "tcp[0]
!= 0 && tcp[1] != 0 && tcp[2] != 0 && tcp[3] != 0 && tcp[4] != 0" for
DLT_EN10MB the very first JEQ instruction instead of "jeq 0x800,
0x00000010, 0x000004b8" was disassembled as "jeq 0x800, 0x00000010,
0xfffffffffffffcb8", the ASCII lines and the dot graph were incorrect.

Fix that by switching to unsigned values.