Handle aarch64 XMM register profile ##debug
When using aarch64 QEMU I had failing assertions in my terminal.
I inspected the output of the register profile and found that it
was failing on:

	xmm	v0	.128	.2144	0
	xmm	v1	.128	.2272	0
	xmm	v2	.128	.2400	0
	xmm	v3	.128	.2528	0

The previous entries, which are `gpr`, are handled. Instead of adding
`xmm` back to the table, we can instead convert it into the `vec128`
counterpart. I wonder which other entries are missing.