Fix iS. paddr vaddr confusion ##core
This change removes the part of the `printHere` address check where it
explicitly checked for `paddr` even if the paddr / vaddr dichotomy is
already handled via the `addr` variable, which is the only one we have
to check here.

That caused wrong behaviour in binaries where the vaddr and paddr
space numerically overlap (like old DOS/4GW LE binaries) where `iS.`
was in fact picking the `paddr` instead of `vaddr` to filter if such
an entry came first in the list.