Fix handling of LE cross-page relocs ##bin
This change fixes an issue for which whenever we encountered negative
`source` offset we gave up on the entire page missing out the rest of
the relocs on that page.

Instead negative means it is a cross-page fixup which is defined in
both pages (started N bytes before the beginning of the current page),
and we can also use this fact to avoid dupe relocs.