Rebase LE objects independently ##bin
This change introduces the optional environment variable `R_LE_BASES`
which can contain a `:`-separated list of addresses where the base of
each defined object should be. This can be used to easily match the
same memory layout as a running DOS system can have while debugging.

The env variable is not great, but it's needed since each object is
rebased independently, instead of "rigidly translated" like is more
common on modern systems where the existing concept of a single base
address would suffice.