Fix dyldcache deps carving logic ##bin
The dyldcache-specific logic which parses library-loading Mach-O load
commands to carve dependencies in order to filter them was wrongly
assuming the path of the loaded library is always at offset 24 instead
of looking up the right offset from the command itself.

While this was working 99.99% of the times, when it didn't it broke the
deps carving logic and produced weird-looking "alien dep" errors.