Add iP command to dsc to get authenticated pointer info ##io
Add the iP command to the dsc io plugin, to get info about a Pointer. That draws information from the underlying rebase infos which are embedded in the pointer itself (whose semantics potentially changes over different maps), which otherwise gets lost as part of rebasing / pointer cleanup.

This is especially useful for getting info about pointer authentication, like the diversity and the key to use if there's any need to ever re-sign a pointer from the dyld cache.

The usage is :iP [size][@vaddr] if no size is provided defaults to 8, runs at current seek (or virtual seek if provided). The output is JSON only.

Example:

```
# this is how we see the pointer stored there normally
> pxq 8@0x1f242f210
0x1f242f210  0x00000001b972985c                       \.r.....

# and this digs the hidden metadata
> :iP @0x1f242f210~{}
[
  {
    "paddr": "0x707b7210",
    "raw": "0x8009b5883972985c",
    "format": "v3",
    "has_diversity": true,
    "diversity": 46472,
    "key": "ia"
  }
]
```

Experimental in the sense that it's quite naive and there's no guard against the user pointing it to an unaligned address for example, where the bits will be interpreted in the wrong way.