Fix DWARF5 file parsing ##dwarf
Due a wrong reinitialization of count, files were all written in the 0'th index,
saving only the last one found.

In DWARF5 the 0'th index is the compilation unit directory which others might be
relative to. This is a problem because olders DWARFs indexes start at 1, and
current code translates 1 to 0 as usual for indexing C arrays, but then when a
set_file opcode sets the file to 0 (clang-18 emits this, gcc-14 do not), the
code can't add a the file because it would underflow the array.

Fix #23579

Signed-off-by: wagner riffel <w@104d.net>