Fixing ext3 detection while detecting filesystem details
In the Extended File Systems, the version is not stored in a
dedicated single "version" field that says "This is ext2/ext3/ext4."
Instead, the version is implied through a combination of feature
flags stored in the superblock.

I made a mistake while checking if the filesystem has journal, the
main difference between ext2 and ext3.